load-tests/requirements.txt
locust>=2.20.0
python-dotenv>=1.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
faker>=20.0.0
msal>=1.24.0
